Add DevTestLab 2021-09-01 API definition Microsoft.DevTestLab supports a variety of new calls and resource types with API Version 2021-09-01 Add the definition to the rest api spec repository for SDK generation.
This is the first new API version since 2018-10-15-preview
Besides new actions and resource types, the update also attempts to align naming and object definitions with ARM expectations.
Previous versions of the API exposes properties in get/patch operations that aren't actually supported by the target resource types.

As a result, there will be a number of breaking changes to the definitions; however these remove properties weren't being used on the backend with prior API versions.